hlalds

+++ philo/utils.c
@@ -6,7 +6,7 @@
 /*   By: alcaball <alcaball@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/28 12:49:21 by alcaball          #+#    #+#             */
-/*   Updated: 2023/11/07 12:18:12 by alcaball         ###   ########.fr       */
+/*   Updated: 2023/11/11 11:26:42 by alcaball         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,5 +43,5 @@ unsigned long	my_time(void)
 
 	if (gettimeofday(&abstime, NULL) == -1)
 		return (write(2, "ERROR getting time", 18));
-	return (abstime.tv_sec * 1000 + abstime.tv_usec / 1000);
+	return ((abstime.tv_sec * (u_int64_t)1000) + (abstime.tv_usec / 1000));
 }
